Create LoginProps interface define the menu props
The Login button that resides in the navbar has the menu props to
identify that it is placed in the navbar. The Login button that
is not in the navbar doesn't have the menu props. This way, we can
style the Login button differently depending if it is in the navbar
or not.